requirements.txt
aiohttp>=3.9.5
aiofiles>=23.2.1
annotated-types>=0.7.0
anyio>=4.3.0
certifi>=2025.7.14
colorama>=0.4.6
distro>=1.9.0
fastapi>=0.111.0
google-cloud-texttospeech>=2.16.2
gTTS>=2.5.4
h11>=0.14.0
httpcore>=1.0.5
httpx>=0.28.1
idna>=3.7
jiter>=0.10.0
openai>=1.97.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv>=1.1.1
python-telegram-bot[webhooks]>=22.3
requests>=2.32.4
six>=1.17.0
sniffio>=1.3.1
tqdm>=4.67.1
typing-inspection>=0.4.1
typing_extensions>=4.14.1
urllib3>=2.5.0
uvicorn[standard]>=0.30.1
